Tidy up code and namespaces
